Introduce piano and drum kit modes.
They are the shift functions of the sixth and seventh scene buttons respectively, and
turn the 8x8 button grid into playable midi instruments.

In these modes
* the scene buttons still launch scenes
* WIP: the track buttons now set midi note velocity
* the up/down bank buttons control the playable pitch range
* the left/right bank buttons control output channel.

To avoid mapping conflicts:
* UI signals are sent on channel 0 only
* Piano outputs are restricted to channels 1 to 11
* Drum kit outputs are restricted to channels 12 to 15

The piano instrument has two layout modes, which can be selected by repeatedly
shift-pressing the the sixth scene button.

Intially the piano is in the "full" layout:
* Each octave is displayed as 12 semitones
* The root of the octave is shown in green
* Middle C is shown in red
* The "white" keys (in a c major scale) are lit amber
* Remaining keys are unlit

The other layout is "brief":
* Each octave is displayed as 7 tones and the succeeding octave
* The root of the octave is shown in green
* Middle C is shown in red
* The major 3rd and perfect fifth are lit amber
* Remaining keys are unlit